requirements.txt
alembic==1.4.2
awscli==1.18.97
Babel==2.8.0
Click==7.0
colorama==0.4.3
cryptography==2.9.2
ecdsa==0.15
Flask==1.0.3
Flask-Cors==3.0.7
Flask-Migrate==2.5.3
Flask-Moment==0.9.0
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.3
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.1.2
MarkupSafe==1.1.1
mccabe==0.6.1
psycopg2==2.8.5
psycopg2-binary==2.8.5
PyJWT==1.7.1
python-dateutil==2.6.0
python-editor==1.0.4
python-jose==3.1.0
six==1.12.0
SQLAlchemy==1.3.4
Werkzeug==1.0.1